changed classnames to be hyphenated, rewrote som modules to use SpaceTradersGet class and added view systems module
