+++ test/view-ship.test.js
@@ -5,7 +5,7 @@ document.getElementById('mochaTestButton').addEventListener('click', () => mocha
 describe('View Ship module', function() {
     it('should return an [HTTP 200 OK] status code', async function() {
 
-        let viewShipResponseStatus = localStorage.getItem('viewShipResponse');
+        let viewShipResponseStatus = localStorage.getItem('view-ship-response.status');
         assert.equal(viewShipResponseStatus, 'true');
     });
 });
